Store filter in top-level state; add tests for filtering
We now store the filter in top-level state. We also store the list of
tabs that match this filter. When the filter changes, we no longer
add/remove classes to tabs. Instead, we do not render tabs that are
filtered out.

Filters now support case-insensitive matching. Newly created and updated
tabs now will respect the current filter.